+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from "./lib/util";
-export * from "./types";
+export * from "./lib/util.js";
+export * from "./types.js";